+++ tests/test_column_loading.py
@@ -1,8 +1,12 @@
 from unittest import TestCase
 
+from packaging.version import Version
+import sqlalchemy as sa
 from sqlalchemy.types import NullType, VARCHAR
 
 from sqlalchemy_redshift.dialect import RedshiftDialect
+
+sa_version = Version(sa.__version__)
 
 
 class TestColumnReflection(TestCase):
@@ -11,13 +15,31 @@ class TestColumnReflection(TestCase):
         Varchar columns with no length should be considered NullType columns
         """
         dialect = RedshiftDialect()
-        column_info = dialect._get_column_info(
-            'Null Column',
-            'character varying', None, False, {}, {}, 'default'
+
+        null_info = dialect._get_column_info(
+            name='Null Column',
+            format_type='character varying',
+            default=None,
+            notnull=False,
+            domains={},
+            enums=[],
+            schema='default',
+            encode='',
+            comment='test column',
+            identity=None
         )
-        assert isinstance(column_info['type'], NullType)
-        column_info_1 = dialect._get_column_info(
-            'character column',
-            'character varying(30)', None, False, {}, {}, 'default'
+        assert isinstance(null_info['type'], NullType)
+
+        varchar_info = dialect._get_column_info(
+            name='character column',
+            format_type='character varying(30)',
+            default=None,
+            notnull=False,
+            domains={},
+            enums=[],
+            schema='default',
+            encode='',
+            comment='test column',
+            identity=None
         )
-        assert isinstance(column_info_1['type'], VARCHAR)
+        assert isinstance(varchar_info['type'], VARCHAR)