+++ tests/test_copy_command.py
@@ -3,7 +3,6 @@ import sqlalchemy as sa
 from sqlalchemy import exc as sa_exc
 
 from sqlalchemy_redshift import dialect
-from sqlalchemy_redshift import commands
 from rs_sqla_test_utils.utils import clean, compile_query
 
 access_key_id = 'IO1IWSZL5YRFM3BEW256'
@@ -79,6 +78,91 @@ def test_iam_role():
     assert clean(expected_result) == clean(compile_query(copy))
 
 
+def test_iam_role_partition():
+    """Tests the use of iam role with a custom partition"""
+
+    aws_partition = 'aws-us-gov'
+    aws_account_id = '000123456789'
+    iam_role_name = 'redshiftrole'
+    creds = 'aws_iam_role=arn:{0}:iam::{1}:role/{2}'.format(
+        aws_partition,
+        aws_account_id,
+        iam_role_name,
+    )
+
+    expected_result = """
+    COPY schema1.t1 FROM 's3://mybucket/data/listing/'
+    WITH CREDENTIALS AS '{creds}'
+    """.format(creds=creds)
+
+    copy = dialect.CopyCommand(
+        tbl,
+        data_location='s3://mybucket/data/listing/',
+        aws_partition=aws_partition,
+        aws_account_id=aws_account_id,
+        iam_role_name=iam_role_name,
+    )
+    assert clean(expected_result) == clean(compile_query(copy))
+
+
+def test_iam_role_partition_validation():
+    """Tests the use of iam role with an invalid partition"""
+
+    aws_partition = 'aws-invalid'
+    aws_account_id = '000123456789'
+    iam_role_name = 'redshiftrole'
+    with pytest.raises(ValueError):
+        dialect.CopyCommand(
+            tbl,
+            data_location='s3://mybucket/data/listing/',
+            aws_partition=aws_partition,
+            aws_account_id=aws_account_id,
+            iam_role_name=iam_role_name,
+        )
+
+
+def test_iam_role_arns_list():
+    """Tests the use of multiple iam role arns instead of access keys."""
+
+    iam_role_arns = [
+        'arn:aws:iam::000123456789:role/redshiftrole',
+        'arn:aws:iam::000123456789:role/redshiftrole2',
+    ]
+    creds = 'aws_iam_role=arn:aws:iam::000123456789:role/redshiftrole,' \
+            'arn:aws:iam::000123456789:role/redshiftrole2'
+
+    expected_result = """
+    COPY schema1.t1 FROM 's3://mybucket/data/listing/'
+    WITH CREDENTIALS AS '{creds}'
+    """.format(creds=creds)
+
+    copy = dialect.CopyCommand(
+        tbl,
+        data_location='s3://mybucket/data/listing/',
+        iam_role_arns=iam_role_arns,
+    )
+    assert clean(expected_result) == clean(compile_query(copy))
+
+
+def test_iam_role_arns_single():
+    """Tests the use of a single iam role arn instead of access keys."""
+
+    iam_role_arns = 'arn:aws:iam::000123456789:role/redshiftrole'
+    creds = 'aws_iam_role=arn:aws:iam::000123456789:role/redshiftrole'
+
+    expected_result = """
+    COPY schema1.t1 FROM 's3://mybucket/data/listing/'
+    WITH CREDENTIALS AS '{creds}'
+    """.format(creds=creds)
+
+    copy = dialect.CopyCommand(
+        tbl,
+        data_location='s3://mybucket/data/listing/',
+        iam_role_arns=iam_role_arns,
+    )
+    assert clean(expected_result) == clean(compile_query(copy))
+
+
 def test_format():
     expected_result = """
     COPY t1 FROM 's3://mybucket/data/listing/'
@@ -106,8 +190,8 @@ def test_format():
 
 
 @pytest.mark.parametrize('format_type', (
-    commands.Format.orc,
-    commands.Format.parquet,
+    dialect.Format.orc,
+    dialect.Format.parquet,
 ))
 def test_format__columnar(format_type):
     expected_result = """
@@ -140,7 +224,7 @@ def test_invalid_format():
 def test_fixed_width_format_without_widths():
     copy = dialect.CopyCommand(
         tbl,
-        format=commands.Format.fixed_width,
+        format=dialect.Format.fixed_width,
         data_location='s3://bucket',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key