+++ tests/test_unload_from_select.py
@@ -1,6 +1,6 @@
+import pytest
 import sqlalchemy as sa
 
-from sqlalchemy_redshift import commands
 from sqlalchemy_redshift import dialect
 
 from rs_sqla_test_utils.utils import clean, compile_query
@@ -60,6 +60,98 @@ def test_iam_role():
         unload_location='s3://bucket/key',
         aws_account_id=aws_account_id,
         iam_role_name=iam_role_name,
+    )
+
+    expected_result = """
+        UNLOAD ('SELECT count(t1.id) AS count_1 FROM t1')
+        TO 's3://bucket/key'
+        CREDENTIALS '{creds}'
+    """.format(creds=creds)
+
+    assert clean(compile_query(unload)) == clean(expected_result)
+
+
+def test_iam_role_partition():
+    """Tests the use of iam role with a custom partition"""
+
+    aws_partition = 'aws-us-gov'
+    aws_account_id = '000123456789'
+    iam_role_name = 'redshiftrole'
+    creds = 'aws_iam_role=arn:{0}:iam::{1}:role/{2}'.format(
+        aws_partition,
+        aws_account_id,
+        iam_role_name,
+    )
+
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        aws_partition=aws_partition,
+        aws_account_id=aws_account_id,
+        iam_role_name=iam_role_name,
+    )
+
+    expected_result = """
+        UNLOAD ('SELECT count(t1.id) AS count_1 FROM t1')
+        TO 's3://bucket/key'
+        CREDENTIALS '{creds}'
+    """.format(creds=creds)
+
+    assert clean(compile_query(unload)) == clean(expected_result)
+
+
+def test_iam_role_partition_validation():
+    """Tests the use of iam role with an invalid partition"""
+
+    aws_partition = 'aws-invalid'
+    aws_account_id = '000123456789'
+    iam_role_name = 'redshiftrole'
+
+    with pytest.raises(ValueError):
+        dialect.UnloadFromSelect(
+            select=sa.select([sa.func.count(table.c.id)]),
+            unload_location='s3://bucket/key',
+            aws_partition=aws_partition,
+            aws_account_id=aws_account_id,
+            iam_role_name=iam_role_name,
+        )
+
+
+def test_iam_role_arns_list():
+    """Tests the use of multiple iam role arns instead of access keys."""
+
+    iam_role_arns = [
+        'arn:aws:iam::000123456789:role/redshiftrole',
+        'arn:aws:iam::000123456789:role/redshiftrole2',
+    ]
+    creds = 'aws_iam_role=arn:aws:iam::000123456789:role/redshiftrole,' \
+            'arn:aws:iam::000123456789:role/redshiftrole2'
+
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        iam_role_arns=iam_role_arns,
+    )
+
+    expected_result = """
+        UNLOAD ('SELECT count(t1.id) AS count_1 FROM t1')
+        TO 's3://bucket/key'
+        CREDENTIALS '{creds}'
+    """.format(creds=creds)
+
+    assert clean(compile_query(unload)) == clean(expected_result)
+
+
+def test_iam_role_arns_single():
+    """Tests the use of a single iam role arn instead of access keys."""
+
+    iam_role_arns = 'arn:aws:iam::000123456789:role/redshiftrole'
+    creds = 'aws_iam_role=arn:aws:iam::000123456789:role/redshiftrole'
+
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        iam_role_arns=iam_role_arns,
     )
 
     expected_result = """
@@ -157,7 +249,7 @@ def test_all_redshift_options_with_header():
     assert clean(compile_query(unload)) == clean(expected_result)
 
 
-def test_csv_format():
+def test_csv_format__basic():
     """Tests that UnloadFromSelect uses the format option correctly."""
 
     unload = dialect.UnloadFromSelect(
@@ -165,7 +257,7 @@ def test_csv_format():
         unload_location='s3://bucket/key',
         access_key_id=access_key_id,
         secret_access_key=secret_access_key,
-        format=commands.Format.csv
+        format=dialect.Format.csv
     )
 
     expected_result = """
@@ -176,3 +268,69 @@ def test_csv_format():
         """.format(creds=creds)
 
     assert clean(compile_query(unload)) == clean(expected_result)
+
+
+@pytest.mark.parametrize('delimiter,fixed_width', (
+    ('\t', None),
+    (None, 'id:8,name:32'),
+    (';', 'id:8,name:32'),
+))
+def test_csv_format__bad_options_crash(delimiter, fixed_width):
+    """Test that UnloadFromSelect crashes if you try to use DELIMITER and/or
+    FIXEDWIDTH with the CSV format.
+    """
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        access_key_id=access_key_id,
+        secret_access_key=secret_access_key,
+        format=dialect.Format.csv,
+        delimiter=delimiter,
+        fixed_width=fixed_width
+    )
+
+    with pytest.raises(ValueError):
+        compile_query(unload)
+
+
+def test_parquet_format__basic():
+    """Basic successful test of unloading with the Parquet format."""
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        access_key_id=access_key_id,
+        secret_access_key=secret_access_key,
+        format=dialect.Format.parquet,
+    )
+
+    expected_result = """
+        UNLOAD ('SELECT count(t1.id) AS count_1 FROM t1')
+        TO 's3://bucket/key'
+        CREDENTIALS '{creds}'
+        FORMAT AS PARQUET
+    """.format(creds=creds)
+
+    assert clean(compile_query(unload)) == clean(expected_result)
+
+
+@pytest.mark.parametrize('kwargs', (
+    {'delimiter': '\t'},
+    {'fixed_width': 'id:8,name:32'},
+    {'gzip': True},
+    {'add_quotes': True, 'escape': True},
+    {'null': '\\N'},
+    {'header': True},
+))
+def test_parquet_format__bad_options_crash(kwargs):
+    """Verify we crash if we use the Parquet format with a bad option."""
+    unload = dialect.UnloadFromSelect(
+        select=sa.select([sa.func.count(table.c.id)]),
+        unload_location='s3://bucket/key',
+        access_key_id=access_key_id,
+        secret_access_key=secret_access_key,
+        format=dialect.Format.parquet,
+        **kwargs
+    )
+
+    with pytest.raises(ValueError):
+        compile_query(unload)